requirements.txt
timm==0.9.16
Pillow
blobfile
mypy
pytest
requests
einops
scipy
pyhealth
h5py
mne==1.4.2
tensorboardX
numpy==1.26.4
pandas==2.2.3
wfdb
tqdm
vitaldb
openpyxl
dotmap
rotary_embedding_torch